Created common Tokenization function_pass
