+++ src/common/snippets/include/snippets/pass/collapse_subgraph.hpp
@@ -12,28 +12,6 @@
 namespace ngraph {
 namespace snippets {
 namespace pass {
-/*
- NotSet - default value returned by GetSnippetsNodeType(...) if the node wasn't marked
- SkippedByPlugin - indicate that snippets can't include this node in subgraph. Can be set by Plugin via SetSnippetsNodeType(...).
- */
-enum class SnippetsNodeType : int64_t {NotSet, SkippedByPlugin};
-void SetSnippetsNodeType(const std::shared_ptr<Node>&, SnippetsNodeType);
-SnippetsNodeType GetSnippetsNodeType(const std::shared_ptr<const Node>&);
-void SetTopologicalOrder(const std::shared_ptr<Node>&, int64_t);
-int64_t GetTopologicalOrder(const std::shared_ptr<const Node>&);
-bool AppropriateForSubgraph(const std::shared_ptr<const Node>&);
-
-/**
- * @interface EnumerateNodes
- * @brief  Snippets rely on topological order to avoid creating cyclic dependencies. This transformation sets the order.
- * @ingroup snippets
- */
-class EnumerateNodes : public ov::pass::ModelPass {
-public:
-    OPENVINO_RTTI("EnumerateNodes", "0");
-    EnumerateNodes() : ModelPass() {}
-    bool run_on_model(const std::shared_ptr<ov::Model>&) override;
-};
 
 /**
  * @interface TokenizeSnippets
@@ -61,6 +39,8 @@ class TokenizeSnippets: public ngraph::pass::MatcherPass {
 public:
     OPENVINO_RTTI("TokenizeSnippets", "0");
     explicit TokenizeSnippets();
+
+    static bool AppropriateForSubgraph(const std::shared_ptr<const Node>&);
 
     static const std::set<ngraph::element::Type> supported_element_types;
 };
