+++ src/common/snippets/tests/src/pass/mha_tokenization.cpp
@@ -5,8 +5,7 @@
 #include <gtest/gtest.h>
 #include <pass/mha_tokenization.hpp>
 #include <subgraph_mha.hpp>
-#include "snippets/pass/collapse_subgraph.hpp"
-#include "snippets/pass/mha_tokenization.hpp"
+#include "snippets/pass/tokenization.hpp"
 #include "snippets/pass/explicit_transpose_matmul_inputs.hpp"
 
 namespace ov {
