+++ src/plugins/intel_cpu/tests/unit/snippets/fake_quantize_tokenization_test.cpp
@@ -6,7 +6,7 @@
 
 #include "common_test_utils/ngraph_test_utils.hpp"
 #include "snippets/pass/fq_decomposition.hpp"
-#include "snippets/pass/collapse_subgraph.hpp"
+#include "snippets/pass/tokenization.hpp"
 #include "fake_quantize_function.hpp"
 #include "snippets/op/subgraph.hpp"
 #include "ngraph_transformations/snippets_mark_skipped.hpp"
